backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.36.0
python-dotenv==1.1.1
pydantic[email]==2.11.9
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
faiss-cpu==1.12.0
langchain==0.3.27
langchain-openai==0.3.33
langchain-community==0.3.29
